+++ include/mujs/mujs.h
@@ -245,6 +245,10 @@ void js_loadstringE(js_State *J, const char *filename, const char *source);
 #define JS_MAX_INT +9007199254740992LL
 #define JS_MIN_INT -9007199254740992LL
 #define JS_MAX_STRING 0x7fffffff
+/* object bound registry */
+void js_getlocalregistry(js_State *J, int idx, const char *name);
+void js_setlocalregistry(js_State *J, int idx, const char *name);
+void js_dellocalregistry(js_State *J, int idx, const char *name);
 
 #ifdef __cplusplus
 }